requirements.txt
Django==4.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
mysql==0.0.3
mysqlclient==2.1.1
requests==2.28.0